+++ zarrs_metadata/src/array/codec/ext/shuffle.rs
@@ -1,0 +1,45 @@
+use derive_more::{Display, From};
+use serde::{Deserialize, Serialize};
+
+use crate::v3::MetadataConfigurationSerialize;
+
+/// The identifier for the `shuffle` codec.
+pub const IDENTIFIER: &str = "shuffle";
+
+/// A wrapper to handle various versions of `shuffle` codec configuration parameters.
+#[derive(Serialize, Deserialize, Clone, PartialEq, Debug, Display, From)]
+#[non_exhaustive]
+#[serde(untagged)]
+pub enum ShuffleCodecConfiguration {
+    /// Version 1.0 draft.
+    V1(ShuffleCodecConfigurationV1),
+}
+
+impl MetadataConfigurationSerialize for ShuffleCodecConfiguration {}
+
+/// `shuffle` codec configuration parameters.
+#[derive(Serialize, Deserialize, Clone, PartialEq, Debug, Display)]
+#[serde(deny_unknown_fields)]
+#[display("{}", serde_json::to_string(self).unwrap_or_default())]
+pub struct ShuffleCodecConfigurationV1 {
+    /// The element size.
+    pub elementsize: usize,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn shuffle() {
+        let configuration = serde_json::from_str::<ShuffleCodecConfigurationV1>(
+            r#"
+        {
+            "elementsize": 4
+        }
+        "#,
+        )
+        .unwrap();
+        assert_eq!(configuration.elementsize, 4);
+    }
+}